refactor - code composition (removed repetitive code)
