+++ src/pages/Bookmark.tsx
@@ -1,9 +1,5 @@
-import { Box } from "../utilities/material-ui/material-components";
 import {
-  QuestionCardsSection,
-  TrendingSection,
-  AddPostMobile,
-  Sidenav,
+  QuestionCardsSection
 } from "../components/index";
 import React from "react";
 import { useAppDispatch, useAppSelector } from "../redux/customHook";
@@ -13,8 +9,6 @@ import { Question } from "../types/Question";
 import { IDLE } from "../utilities/constants/api-status";
 
 export function Bookmark() {
-  const [open, setOpen] = React.useState(false);
-  const handleOpen = () => setOpen(true);
   const dispatch = useAppDispatch();
   const { questions, questionStatus } = useAppSelector(
     (state: RootState) => state.feed
@@ -35,22 +29,9 @@ export function Bookmark() {
   }, [questions]);
 
   return (
-    <Box
-      sx={{
-        display: "flex",
-        p: { md: "0 1rem 0 0", xs: "1rem" },
-        gap: 2,
-        flexDirection: { xs: "column-reverse", md: "row" },
-      }}
-      component="main"
-    >
-      <Sidenav handleOpen={handleOpen} open={open} setOpen={setOpen} />
-      <QuestionCardsSection
-        title="Your Bookmarks"
-        questions={questionsToDisplay}
-      />
-      <TrendingSection />
-      <AddPostMobile handleOpen={handleOpen} />
-    </Box>
+    <QuestionCardsSection
+      title="Your Bookmarks"
+      questions={questionsToDisplay}
+    />
   );
 }
