+++ src/pages/feed.tsx
@@ -1,10 +1,4 @@
-import { Box } from "../utilities/material-ui/material-components";
-import {
-  QuestionCardsSection,
-  TrendingSection,
-  AddPostMobile,
-  Sidenav,
-} from "../components/index";
+import { QuestionCardsSection } from "../components/index";
 import React from "react";
 import { useAppDispatch, useAppSelector } from "../redux/customHook";
 import { getQuestions } from "../redux/slices/feedSlice";
@@ -13,8 +7,6 @@ import { Question } from "../types/Question";
 import { IDLE } from "../utilities/constants/api-status";
 
 export function Feed() {
-  const [open, setOpen] = React.useState(false);
-  const handleOpen = () => setOpen(true);
   const dispatch = useAppDispatch();
   const { questions, questionStatus } = useAppSelector(
     (state: RootState) => state.feed
@@ -67,26 +59,13 @@ export function Feed() {
   }, [questions, tagFilters, sortBy]);
 
   return (
-    <Box
-      sx={{
-        display: "flex",
-        p: { md: "0 1rem 0 0", xs: "1rem" },
-        gap: 2,
-        flexDirection: { xs: "column-reverse", md: "row" },
-      }}
-      component="main"
-    >
-      <Sidenav handleOpen={handleOpen} open={open} setOpen={setOpen} />
-      <QuestionCardsSection
-        title="Latest Posts"
-        questions={questionsToDisplay}
-        tagFilters={tagFilters}
-        updateFilter={updateFilter}
-        sortBy={sortBy}
-        updateSortBy={updateSortBy}
-      />
-      <TrendingSection />
-      <AddPostMobile handleOpen={handleOpen} />
-    </Box>
+    <QuestionCardsSection
+      title="Latest Posts"
+      questions={questionsToDisplay}
+      tagFilters={tagFilters}
+      updateFilter={updateFilter}
+      sortBy={sortBy}
+      updateSortBy={updateSortBy}
+    />
   );
 }
